mqtt-redis/requirements.txt
redis
python-dotenv
paho_mqtt